new_aggregator_/requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.6.0
pymongo==3.11.0
pytz==2020.4
requests==2.24.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
